Network.HTTP.Client.Conduit

Description

A new, experimental API to replace Network.HTTP.Conduit.

For most users, Network.HTTP.Simple is probably a better choice. For more information, see:

https://haskell-lang.org/library/http-client

For more information on using this module, please be sure to read the documentation in the Network.HTTP.Client module.

Conduit-specific interface

withResponse :: (MonadUnliftIO m, MonadIO n, MonadReader env m, HasHttpManager env) => Request -> (Response (ConduitM i ByteString n ()) -> m a) -> m a

Conduit powered version of withResponse. Differences are:

Response body is represented as a Producer.

Generalized to any instance of MonadUnliftIO, not just IO.

The Manager is contained by a MonadReader context.

Since 2.1.0

responseOpen :: (MonadIO m, MonadIO n, MonadReader env m, HasHttpManager env) => Request -> m (Response (ConduitM i ByteString n ()))

Conduit-powered version of responseOpen.

See withResponse for the differences with responseOpen.

Since 2.1.0

responseClose :: MonadIO m => Response body -> m ()

Generalized version of responseClose.

Since 2.1.0

acquireResponse :: (MonadIO n, MonadReader env m, HasHttpManager env) => Request -> m (Acquire (Response (ConduitM i ByteString n ())))

An Acquire for getting a Response.

Since 2.1.0

httpSource :: (MonadResource m, MonadIO n, MonadReader env m, HasHttpManager env) => Request -> (Response (ConduitM () ByteString n ()) -> ConduitM () r m ()) -> ConduitM () r m ()

Same as httpSource, but uses Manager from Reader environment instead of the global one.

Since 2.3.6

Manager helpers

defaultManagerSettings :: ManagerSettings

TLS-powered manager settings.

Since 2.1.0

newManager :: MonadIO m => m Manager

Get a new manager using defaultManagerSettings.

Since 2.1.0

newManagerSettings :: MonadIO m => ManagerSettings -> m Manager

Get a new manager using the given settings.

Since 2.1.0

General HTTP client interface

responseTimeoutDefault :: ResponseTimeout

Use the default response timeout

When used on a Request, means: use the manager's timeout value

When used on a ManagerSettings, means: default to 30 seconds

Since: http-client-0.5.0

responseTimeoutNone :: ResponseTimeout

Do not have a response timeout

Since: http-client-0.5.0

responseTimeoutMicro :: Int -> ResponseTimeout

Specify maximum time in microseconds the retrieval of response headers is allowed to take

Since: http-client-0.5.0

managerSetProxy :: ProxyOverride -> ManagerSettings -> ManagerSettings

Set the proxy override value, for both HTTP (insecure) and HTTPS (insecure) connections.

Since 0.4.7

managerSetSecureProxy :: ProxyOverride -> ManagerSettings -> ManagerSettings

Set the proxy override value, only for HTTPS (secure) connections.

Since 0.4.7

managerSetInsecureProxy :: ProxyOverride -> ManagerSettings -> ManagerSettings

Set the proxy override value, only for HTTP (insecure) connections.

Since 0.4.7

withResponseHistory :: Request -> Manager -> (HistoriedResponse BodyReader -> IO a) -> IO a

A variant of withResponse which keeps a history of all redirects performed in the interim, together with the first 1024 bytes of their response bodies.

Since 0.4.1

responseOpenHistory :: Request -> Manager -> IO (HistoriedResponse BodyReader)

A variant of responseOpen which keeps a history of all redirects performed in the interim, together with the first 1024 bytes of their response bodies.

Since 0.4.1

data HistoriedResponse body

A datatype holding information on redirected requests and the final response.

Since 0.4.1

withConnection :: Request -> Manager -> (Connection -> IO a) -> IO a

Perform an action using a Connection acquired from the given Manager.

You should use this only when you have to read and write interactively through the connection (e.g. connection by the WebSocket protocol).

Since: http-client-0.5.13

generateCookie

Arguments

:: SetCookie
The SetCookie we are encountering

-> Request
The request that originated the response that yielded the SetCookie

-> UTCTime
Value that should be used as "now"

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> Maybe Cookie
The optional output cookie

Turn a SetCookie into a Cookie, if it is valid

insertCheckedCookie

Arguments

:: Cookie
The SetCookie the cookie jar is receiving

-> CookieJar
Input cookie jar to modify

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> CookieJar
Updated (or not) cookie jar

Insert a cookie created by generateCookie into the cookie jar (or not if it shouldn't be allowed in)

receiveSetCookie

Arguments

:: SetCookie
The SetCookie the cookie jar is receiving

-> Request
The request that originated the response that yielded the SetCookie

-> UTCTime
Value that should be used as "now"

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> CookieJar
Input cookie jar to modify

-> CookieJar
Updated cookie jar

This corresponds to the algorithm described in Section 5.3 "Storage Model" This function consists of calling generateCookie followed by insertCheckedCookie. Use this function if you plan to do both in a row. generateCookie and insertCheckedCookie are only provided for more fine-grained control.

updateCookieJar

Arguments

:: Response a
Response received from server

-> Request
Request which generated the response

-> UTCTime
Value that should be used as "now"

-> CookieJar
Current cookie jar

-> (CookieJar, Response a)
(Updated cookie jar with cookies from the Response, The response stripped of any "Set-Cookie" header)

This applies receiveSetCookie to a given Response

computeCookieString

Arguments

:: Request
Input request

-> CookieJar
Current cookie jar

-> UTCTime
Value that should be used as "now"

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> (ByteString, CookieJar)
(Contents of a "Cookie" header, Updated cookie jar (last-access-time is updated))

This corresponds to the algorithm described in Section 5.4 "The Cookie Header"

insertCookiesIntoRequest

Arguments

:: Request
The request to insert into

-> CookieJar
Current cookie jar

-> UTCTime
Value that should be used as "now"

-> (Request, CookieJar)
(Output request, Updated cookie jar (last-access-time is updated))

This applies the computeCookieString to a given Request

evictExpiredCookies

Arguments

:: CookieJar
Input cookie jar

-> UTCTime
Value that should be used as "now"

-> CookieJar
Filtered cookie jar

This corresponds to the eviction algorithm described in Section 5.3 "Storage Model"

removeExistingCookieFromCookieJar :: Cookie -> CookieJar -> (Maybe Cookie, CookieJar)

destroyCookieJar :: CookieJar -> [Cookie]

createCookieJar :: [Cookie] -> CookieJar

pathMatches :: ByteString -> ByteString -> Bool

This corresponds to the subcomponent algorithm entitled "Path-Match" detailed in section 5.1.4

defaultPath :: Request -> ByteString

This corresponds to the subcomponent algorithm entitled "Paths" detailed in section 5.1.4

domainMatches

Arguments

:: ByteString
Domain to test

-> ByteString
Domain from a cookie

-> Bool

This corresponds to the subcomponent algorithm entitled "Domain Matching" detailed in section 5.1.3

isIpAddress :: ByteString -> Bool

defaultProxy :: ProxyOverride

The default proxy settings for a manager. In particular: if the http_proxy (or https_proxy) environment variable is set, use it. Otherwise, use the values in the Request.

Since 0.4.7

proxyEnvironmentNamed

Arguments

:: Text
environment variable name

-> Maybe Proxy
fallback if no environment set

-> ProxyOverride

Same as proxyEnvironment, but instead of default environment variable names, allows you to set your own name.

Since 0.4.7

proxyEnvironment

Arguments

:: Maybe Proxy
fallback if no environment set

-> ProxyOverride

Get the proxy settings from the default environment variable (http_proxy for insecure, https_proxy for secure). If no variable is set, then fall back to the given value. Nothing is equivalent to noProxy, Just is equivalent to useProxy.

Since 0.4.7

useProxy :: Proxy -> ProxyOverride

Use the given proxy settings, regardless of the proxy value in the Request.

Since 0.4.7

noProxy :: ProxyOverride

Never connect using a proxy, regardless of the proxy value in the Request.

Since 0.4.7

proxyFromRequest :: ProxyOverride

Get the proxy settings from the Request itself.

Since 0.4.7

withManager :: ManagerSettings -> (Manager -> IO a) -> IO a

Create, use and close a Manager.

Since 0.2.1

rawConnectionModifySocketSize :: (Socket -> IO ()) -> IO (Int -> Maybe HostAddress -> String -> Int -> IO Connection)

Same as rawConnectionModifySocket, but also takes in a chunk size.

Since: http-client-0.5.2

rawConnectionModifySocket :: (Socket -> IO ()) -> IO (Maybe HostAddress -> String -> Int -> IO Connection)

A value for the managerRawConnection setting, but also allows you to modify the underlying Socket to set additional settings. For a motivating use case, see: https://github.com/snoyberg/http-client/issues/71.

Since 0.3.8

observedStreamFile :: (StreamFileStatus -> IO ()) -> FilePath -> IO RequestBody

Send a file as the request body, while observing streaming progress via a PopObserver. Observations are made between reading and sending a chunk.

It is expected that the file size does not change between calling observedStreamFile and making any requests using this request body.

Since 0.4.9

streamFile :: FilePath -> IO RequestBody

Send a file as the request body.

It is expected that the file size does not change between calling streamFile and making any requests using this request body.

Since 0.4.9

setQueryStringPartialEscape :: [(ByteString, [EscapeItem])] -> Request -> Request

Set the query string to the given key/value pairs.

Since: http-client-0.5.10

setQueryString :: [(ByteString, Maybe ByteString)] -> Request -> Request

Set the query string to the given key/value pairs.

Since 0.3.6

setRequestCheckStatus :: Request -> Request

Modify the request so that non-2XX status codes generate a runtime StatusCodeException, by using throwErrorStatusCodes

Since: http-client-0.5.13

setRequestIgnoreStatus :: Request -> Request

Modify the request so that non-2XX status codes do not generate a runtime StatusCodeException.

Since: http-client-0.4.29

urlEncodedBody :: [(ByteString, ByteString)] -> Request -> Request

Add url-encoded parameters to the Request.

This sets a new requestBody, adds a content-type request header and changes the method to POST.

Since 0.1.0

applyBasicProxyAuth :: ByteString -> ByteString -> Request -> Request

Add a Proxy-Authorization header (with the specified username and password) to the given Request. Ignore error handling:

applyBasicProxyAuth "user" "pass" <$> parseRequest "http://example.org"

Since 0.3.4

applyBasicAuth :: ByteString -> ByteString -> Request -> Request

Add a Basic Auth header (with the specified user name and password) to the given Request. Ignore error handling:

 applyBasicAuth "user" "pass" $ parseRequest_ url

NOTE: The function applyDigestAuth is provided by the http-client-tls package instead of this package due to extra dependencies. Please use that package if you need to use digest authentication.

Since 0.1.0

defaultRequest :: Request

A default request value, a GET request of localhost/:80, with an empty request body.

Note that the default checkResponse does nothing.

Since: http-client-0.4.30

getUri :: Request -> URI

Extract a URI from the request.

Since 0.1.0

requestFromURI_ :: URI -> Request

Same as requestFromURI, but if the conversion would fail, throws an impure exception.

Since: http-client-0.5.12

requestFromURI :: MonadThrow m => URI -> m Request

Convert a URI into a Request.

This can fail if the given URI is not absolute, or if the URI scheme is not "http" or "https". In these cases the function will throw an error via MonadThrow.

This function defaults some of the values in Request, such as setting method to GET and requestHeaders to [].

A Request created by this function won't cause exceptions on non-2XX response status codes.

Since: http-client-0.5.12

parseRequest_ :: String -> Request

Same as parseRequest, but parse errors cause an impure exception. Mostly useful for static strings which are known to be correctly formatted.

parseRequest :: MonadThrow m => String -> m Request

Convert a URL into a Request.

This function defaults some of the values in Request, such as setting method to GET and requestHeaders to [].

Since this function uses MonadThrow, the return monad can be anything that is an instance of MonadThrow, such as IO or Maybe.

You can place the request method at the beginning of the URL separated by a space, e.g.:

@@ parseRequest "POST http://httpbin.org/post" @@

Note that the request method must be provided as all capital letters.

A Request created by this function won't cause exceptions on non-2XX response status codes.

To create a request which throws on non-2XX status codes, see parseUrlThrow

Since: http-client-0.4.30

throwErrorStatusCodes :: MonadIO m => Request -> Response BodyReader -> m ()

Throws a StatusCodeException wrapped in HttpExceptionRequest, if the response's status code indicates an error (if it isn't 2xx). This can be used to implement checkResponse.

Since: http-client-0.5.13

parseUrlThrow :: MonadThrow m => String -> m Request

Same as parseRequest, except will throw an HttpException in the event of a non-2XX response. This uses throwErrorStatusCodes to implement checkResponse.

Since: http-client-0.4.30

parseUrl :: MonadThrow m => String -> m Request

Deprecated synonym for parseUrlThrow. You probably want parseRequest or parseRequest_ instead.

Since: http-client-0.1.0

brReadSome :: BodyReader -> Int -> IO ByteString

Continuously call brRead, building up a lazy ByteString until a chunk is constructed that is at least as many bytes as requested.

Since 0.4.20

socketConnection

Arguments

:: Socket

-> Int
chunk size

-> IO Connection

Create a new Connection from a Socket.

Since: http-client-0.5.3

makeConnection

Arguments

:: IO ByteString
read

-> (ByteString -> IO ())
write

-> IO ()
close

-> IO Connection

Create a new Connection from a read, write, and close function.

Since: http-client-0.5.3

data HttpException

An exception which may be generated by this library

Since: http-client-0.5.0

Constructors

HttpExceptionRequest Request HttpExceptionContent
Most exceptions are specific to a Request. Inspect the HttpExceptionContent value for details on what occurred.

Since: http-client-0.5.0

InvalidUrlException String String
A URL (first field) is invalid for a given reason (second argument).

Since: http-client-0.5.0

data HttpExceptionContent

Constructors

StatusCodeException (Response ()) ByteString
Generated by the parseUrlThrow function when the server returns a non-2XX response status code.

May include the beginning of the response body.

Since: http-client-0.5.0

TooManyRedirects [Response ByteString]
The server responded with too many redirects for a request.

Contains the list of encountered responses containing redirects in reverse chronological order; including last redirect, which triggered the exception and was not followed.

Since: http-client-0.5.0

OverlongHeaders
Either too many headers, or too many total bytes in a single header, were returned by the server, and the memory exhaustion protection in this library has kicked in.

Since: http-client-0.5.0

ResponseTimeout
The server took too long to return a response. This can be altered via responseTimeout or managerResponseTimeout.

Since: http-client-0.5.0

ConnectionTimeout
Attempting to connect to the server timed out.

Since: http-client-0.5.0

ConnectionFailure SomeException
An exception occurred when trying to connect to the server.

Since: http-client-0.5.0

InvalidStatusLine ByteString
The status line returned by the server could not be parsed.

Since: http-client-0.5.0

InvalidHeader ByteString
The given response header line could not be parsed

Since: http-client-0.5.0

InvalidRequestHeader ByteString
The given request header is not compliant (e.g. has newlines)

Since: http-client-0.5.14

InternalException SomeException
An exception was raised by an underlying library when performing the request. Most often, this is caused by a failing socket action or a TLS exception.

Since: http-client-0.5.0

ProxyConnectException ByteString Int Status
A non-200 status code was returned when trying to connect to the proxy server on the given host and port.

Since: http-client-0.5.0

NoResponseDataReceived
No response data was received from the server at all. This exception may deserve special handling within the library, since it may indicate that a pipelining has been used, and a connection thought to be open was in fact closed.

Since: http-client-0.5.0

TlsNotSupported
Exception thrown when using a Manager which does not have support for secure connections. Typically, you will want to use tlsManagerSettings from http-client-tls to overcome this.

Since: http-client-0.5.0

WrongRequestBodyStreamSize Word64 Word64
The request body provided did not match the expected size.

Provides the expected and actual size.

Since: http-client-0.4.31

ResponseBodyTooShort Word64 Word64
The returned response body is too short. Provides the expected size and actual size.

Since: http-client-0.5.0

InvalidChunkHeaders
A chunked response body had invalid headers.

Since: http-client-0.5.0

IncompleteHeaders
An incomplete set of response headers were returned.

Since: http-client-0.5.0

InvalidDestinationHost ByteString
The host we tried to connect to is invalid (e.g., an empty string).

HttpZlibException ZlibException
An exception was thrown when inflating a response body.

Since: http-client-0.5.0

InvalidProxyEnvironmentVariable Text Text
Values in the proxy environment variable were invalid. Provides the environment variable name and its value.

Since: http-client-0.5.0

ConnectionClosed
Attempted to use a Connection which was already closed

Since: http-client-0.5.0

InvalidProxySettings Text
Proxy settings are not valid (Windows specific currently) @since 0.5.7

data Cookie

Constructors

Cookie

Fields

cookie_name :: ByteString

cookie_value :: ByteString

cookie_expiry_time :: UTCTime

cookie_domain :: ByteString

cookie_path :: ByteString

cookie_creation_time :: UTCTime

cookie_last_access_time :: UTCTime

cookie_persistent :: Bool

cookie_host_only :: Bool

cookie_secure_only :: Bool

cookie_http_only :: Bool

data CookieJar

data Proxy

Define a HTTP proxy, consisting of a hostname and port number.

Constructors

Proxy

Fields

proxyHost :: ByteString

The host name of the HTTP proxy.

proxyPort :: Int

The port number of the HTTP proxy.

data RequestBody

When using one of the RequestBodyStream / RequestBodyStreamChunked constructors, you must ensure that the GivesPopper can be called multiple times. Usually this is not a problem.

The RequestBodyStreamChunked will send a chunked request body. Note that not all servers support this. Only use RequestBodyStreamChunked if you know the server you're sending to supports chunked request bodies.

Since 0.1.0

Constructors

RequestBodyLBS ByteString

RequestBodyBS ByteString

RequestBodyBuilder Int64 Builder

RequestBodyStream Int64 (GivesPopper ())

RequestBodyStreamChunked (GivesPopper ())

RequestBodyIO (IO RequestBody)
Allows creation of a RequestBody inside the IO monad, which is useful for making easier APIs (like setRequestBodyFile).

Since: http-client-0.4.28

type Popper = IO ByteString

A function which generates successive chunks of a request body, provider a single empty bytestring when no more data is available.

Since 0.1.0

type NeedsPopper a = Popper -> IO a

A function which must be provided with a Popper.

Since 0.1.0

type GivesPopper a = NeedsPopper a -> IO a

A function which will provide a Popper to a NeedsPopper. This seemingly convoluted structure allows for creation of request bodies which allocate scarce resources in an exception safe manner.

Since 0.1.0

data Request

All information on how to connect to a host and what should be sent in the HTTP request.

If you simply wish to download from a URL, see parseRequest.

The constructor for this data type is not exposed. Instead, you should use either the defaultRequest value, or parseRequest to construct from a URL, and then use the records below to make modifications. This approach allows http-client to add configuration options without breaking backwards compatibility.

For example, to construct a POST request, you could do something like:

initReq <- parseRequest "http://www.example.com/path"
let req = initReq
            { method = "POST"
            }

For more information, please see http://www.yesodweb.com/book/settings-types.

Since 0.1.0

data ResponseTimeout

How to deal with timing out on retrieval of response headers.

Since: http-client-0.5.0

data Response body

A simple representation of the HTTP response.

Since 0.1.0

data ManagerSettings

Settings for a Manager. Please use the defaultManagerSettings function and then modify individual settings. For more information, see http://www.yesodweb.com/book/settings-types.

Since 0.1.0

data ProxyOverride

How the HTTP proxy server settings should be discovered.

Since 0.4.7

data Manager

Keeps track of open connections for keep-alive.

If possible, you should share a single Manager between multiple threads and requests.

Since 0.1.0

class HasHttpManager a where

Methods

getHttpManager :: a -> Manager

data StreamFileStatus

Status of streaming a request body from a file.

Since 0.4.9

Constructors

StreamFileStatus

Fields

fileSize :: Int64

readSoFar :: Int64

thisChunkSize :: Int

httpLbs :: (MonadIO m, HasHttpManager env, MonadReader env m) => Request -> m (Response ByteString)

Same as httpLbs, except it uses the Manager in the reader environment.

Since 2.1.1

httpNoBody :: (MonadIO m, HasHttpManager env, MonadReader env m) => Request -> m (Response ())

Same as httpNoBody, except it uses the Manager in the reader environment.

This can be more convenient that using withManager as it avoids the need to specify the base monad for the response body.

Since 2.1.2

Lower-level conduit functions

requestBodySource :: Int64 -> ConduitM () ByteString IO () -> RequestBody

requestBodySourceChunked :: ConduitM () ByteString IO () -> RequestBody

bodyReaderSource :: MonadIO m => BodyReader -> ConduitM i ByteString m ()