Network.HTTP.Conduit.Internal

Documentation

sinkHeaders :: (MonadThrow m, MonadResource m) => Sink ByteString m (Status, [Header])

newline :: Parser ()

parserHeadersFromByteString :: Monad m => ByteString -> m (Either String (Status, [Header]))

parseChunkHeader :: Parser Int

getUri :: Request m' -> URI

Extract a URI from the request.

setUri :: Failure HttpException m => Request m' -> URI -> m (Request m')

Validate a URI, then add it to the request.

setUriRelative :: Failure HttpException m => Request m' -> URI -> m (Request m')

Add a URI to the request. If it is absolute (includes a host name), add it as per setUri; if it is relative, merge it with the existing request.

httpRedirect

Arguments

:: (MonadBaseControl IO m, MonadResource m, Monad m1)

=> Int
redirectCount

-> (Request m1 -> m (Response (ResumableSource m1 ByteString), Maybe (Request m1)))
function which performs a request and returns a response, and possibly another request if there's a redirect.

-> (forall a. m1 a -> m a)
liftResourceT

-> Request m1

-> m (Response (ResumableSource m1 ByteString))

Redirect loop